begin moving to memoized reducer function
