+++ airbyte-webapp/src/views/Connection/CatalogDiffModal/CatalogDiffModal.tsx
@@ -3,12 +3,13 @@ import { FormattedMessage } from "react-intl";
 
 import { Button } from "components";
 
-import { AirbyteCatalog, CatalogDiff, FieldTransform, StreamTransform } from "core/request/AirbyteClient";
+import { AirbyteCatalog, CatalogDiff } from "core/request/AirbyteClient";
 
 import { Modal, ModalFooter } from "../../../components/Modal";
 import styles from "./CatalogDiffModal.module.scss";
 import { DiffSection } from "./components/DiffSection";
 import { FieldSection } from "./components/FieldSection";
+import { getSortedDiff } from "./utils/utils";
 
 interface CatalogDiffModalProps {
   catalogDiff: CatalogDiff;
@@ -16,35 +17,8 @@ interface CatalogDiffModalProps {
   setDiffAcknowledged: Dispatch<SetStateAction<boolean>>;
 }
 
-export type DiffVerb = "new" | "removed" | "changed";
-
-interface SortedDiff<T extends StreamTransform | FieldTransform> {
-  newItems: T[];
-  removedItems: T[];
-  changedItems: T[];
-}
-
-export const diffReducer = <T extends StreamTransform | FieldTransform>(diffArray: T[]): SortedDiff<T> => {
-  const sortedDiff: SortedDiff<T> = { newItems: [], removedItems: [], changedItems: [] };
-
-  diffArray.filter((streamTransform) => {
-    if (streamTransform.transformType.includes("add")) {
-      sortedDiff.newItems.push(streamTransform);
-    }
-
-    if (streamTransform.transformType.includes("remove")) {
-      sortedDiff.removedItems.push(streamTransform);
-    }
-
-    if (streamTransform.transformType.includes("update")) {
-      sortedDiff.changedItems.push(streamTransform);
-    }
-    return sortedDiff;
-  });
-  return sortedDiff;
-};
 export const CatalogDiffModal: React.FC<CatalogDiffModalProps> = ({ catalogDiff, catalog, setDiffAcknowledged }) => {
-  const { newItems, removedItems, changedItems } = diffReducer(catalogDiff.transforms);
+  const { newItems, removedItems, changedItems } = getSortedDiff(catalogDiff.transforms);
 
   return (
     <Modal title={<FormattedMessage id="connection.updateSchema.completed" />} onClose={() => null}>
