+++ airbyte-webapp/src/views/Connection/CatalogDiffModal/components/DiffHeader.tsx
@@ -1,6 +1,6 @@
 import { FormattedMessage } from "react-intl";
 
-import { DiffVerb } from "../CatalogDiffModal";
+import { DiffVerb } from "../utils/types";
 
 export type DiffType = "field" | "stream";
