+++ airbyte-webapp/src/views/Connection/CatalogDiffModal/utils/types.tsx
@@ -1,0 +1,9 @@
+import { FieldTransform, StreamTransform } from "core/request/AirbyteClient";
+
+export type DiffVerb = "new" | "removed" | "changed";
+
+export interface SortedDiff<T extends StreamTransform | FieldTransform> {
+  newItems: T[];
+  removedItems: T[];
+  changedItems: T[];
+}
